fix(websocket): Use polling for initial headers
This change ensures that custom headers, such as `ngrok-skip-browser-warning`, are correctly sent during the initial socket connection. Browsers cannot send these headers on pure WebSocket connections, so 'polling' is used as a transport to allow the headers to be included.

This also adds more specific error logging for connection issues, particularly those related to Ngrok.